servidor django completo/requirements.txt
Django
requests
djangorestframework
mysqlclient
